Use integer timestamps.

+++ SilveranKit/Sources/Common/Actors/storyteller/StorytellerActor.swift
@@ -1897,7 +1897,7 @@ public actor StorytellerActor {
 
         let body: [String: Any] = [
             "locator": encodeLocatorToDict(locator),
-            "timestamp": timestamp,
+            "timestamp": Int64(timestamp),
         ]
 
         do {
